% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{assert}
\alias{assert}
\title{Asserts an expression}
\usage{
assert(expr, msg)
}
\arguments{
\item{expr}{The expression to evaluate.}

\item{msg}{The error message.}
}
\description{
If the expression evaluates to false then the execution stops and the given
message is printed.
}
\examples{
\dontrun{
assert(1 == 2, '1 is not equal to 2')
}
}

