% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/util.R
\name{curry}
\alias{curry}
\title{Curries a function}
\usage{
curry(func, ...)
}
\arguments{
\item{func}{The function to curry.}

\item{...}{The parameters to use (named parameters can be used).}
}
\value{
The curried function.
}
\description{
Curries a function. See http://en.wikipedia.org/wiki/Currying for more
information.
}
\examples{
curry(function(x, y) { x + y }, x = 1)(2) # = 3
curry(log, base = 2)(16) # = 4
}

