% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load.data.file}
\alias{load.data.file}
\title{Loads data from a CSV file}
\usage{
load.data.file(fname, pattern, fields, custom.func = NULL, ...)
}
\arguments{
\item{fname}{CSV file name.}

\item{pattern}{Regular expression used to extract the fields values.}

\item{fields}{Fields that uniquely identify the data from a CSV file.}

\item{custom.func}{Function for post-processing the data. Defaults to NULL.}

\item{...}{Arguments to pass to data.table::fread().}
}
\value{
A data.table object containing the data read from the CSV file.
}
\description{
This function loads data from the given CSV file and it adds some extra
data coming from the fields in the file name. The number of capture groups
in the regular expression (pattern) must match the number of fields.
}
\examples{
\dontrun{
load.data.file('results-1-simple.csv',
               'results-(\\\\d+)-(\\\\w+)\\\\.csv',
               c('ID', 'config'))
}
}

