% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/load-data.R
\name{load.data}
\alias{load.data}
\title{Loads data from multiple CSV files and combines the data into a data.table}
\usage{
load.data(path, pattern, fields, custom.func = NULL)
}
\arguments{
\item{path}{Path where to search for the CSV files.}

\item{pattern}{Regular expression used to search the CSV files and extract
the fields values too.}

\item{fields}{Fields that uniquely identify the data from a CSV file.}

\item{custom.func}{Custom function to be used when processing a single CSV
file. Defaults to NULL.}
}
\value{
A data.table object containing the data read from the CSV files.
}
\description{
This function searches for CSV files matching a pattern in the given path
and combines all the data into a data.table object. The number of capture
groups in the regular expression (pattern) must match the number of fields.
}
\examples{
load.data('.', 'results-(\\\\d+)-(\\\\w+)\\\\.csv', c('ID', 'config'))
}

