% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load-data.R
\name{parse.fname}
\alias{parse.fname}
\title{Parses a file name and returns the values for the requested fields}
\usage{
parse.fname(fname, pattern, fields)
}
\arguments{
\item{fname}{file name.}

\item{pattern}{Regular expression used to extract the fields values.}

\item{fields}{Fields that uniquely identify the data from a file.}
}
\value{
A list containing the field names and their values.
}
\description{
This function parses a file name and matches the name to the given pattern.
Then it extracts the values for the requested files.
}
\examples{
\dontrun{
parse.fname('results-1-simple.csv', 'results-(\\\\d+)-(\\\\w+)\\\\.csv', c('ID', 'config'))
}
}

