% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load.data}
\alias{load.data}
\title{Loads data from multiple CSV files and combines the data into a data.table}
\usage{
load.data(path, pattern, fields, local.func = NULL, global.func = NULL)
}
\arguments{
\item{path}{Path where to search for the CSV files.}

\item{pattern}{Regular expression used to search the CSV files and extract
the fields values too.}

\item{fields}{Fields that uniquely identify the data from a CSV file.}

\item{local.func}{Custom function to be called after reading a single CSV
file. Defaults to NULL.}

\item{global.func}{Custom function to be called after all the data tables
have been merged into a single one. Defaults to NULL.}
}
\value{
A data.table object containing the data read from the CSV files.
}
\description{
This function searches for CSV files matching a pattern in the given path
and combines all the data into a data.table object. The number of capture
groups in the regular expression (pattern) must match the number of fields.
}
\examples{
\dontrun{
load.data('.', 'results-(\\\\d+)-(\\\\w+)\\\\.csv', c('ID', 'config'))
}
}

